# native/application/api/openapi/swagger.yaml
basePath: /
definitions:
...
  github_com_gcc798_lightning_application_api_domain_request.BatchDeleteLoginLogRequest:
    properties:
      ids:
        description: 日志ID列表
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - ids
    type: object
...
  github_com_gcc798_lightning_application_api_domain_request.CleanLoginLogRequest:
    properties:
      days:
        description: 清理多少天之前的日志
        minimum: 1
        type: integer
    required:
    - days
    type: object
...
  github_com_gcc798_lightning_application_api_domain_request.CreateLoginLogRequest:
    properties:
      browser:
        description: 浏览器类型
        type: string
      clientId:
        description: 客户端ID
        type: string
      ipaddr:
        description: 登录IP
        type: string
      loginLocation:
        description: 登录地点
        type: string
      msg:
        description: 提示消息
        type: string
      os:
        description: 操作系统
        type: string
      status:
        description: 登录状态：0成功 1失败
        type: integer
      userName:
        description: 用户名
        type: string
    required:
    - userName
    type: object
...
  github_com_gcc798_lightning_application_api_domain_request.UpdateLoginLogRequest:
    properties:
      browser:
        description: 浏览器类型
        type: string
      clientId:
        description: 客户端ID
        type: string
      id:
        description: 日志ID
        type: integer
      ipaddr:
        description: 登录IP
        type: string
      loginLocation:
        description: 登录地点
        type: string
      msg:
        description: 提示消息
        type: string
      os:
        description: 操作系统
        type: string
      status:
        description: 登录状态：0成功 1失败
        type: integer
      userName:
        description: 用户名
        type: string
    required:
    - id
    type: object
...
  github_com_gcc798_lightning_application_api_domain_response.LoginLogResponse:
    properties:
      browser:
        description: 浏览器类型
        type: string
      clientId:
        description: 客户端ID
        type: string
      id:
        description: 日志ID
        type: integer
      ipaddr:
        description: 登录IP
        type: string
      loginLocation:
        description: 登录地点
        type: string
      loginTime:
        description: 登录时间
        type: string
      msg:
        description: 提示消息
        type: string
      os:
        description: 操作系统
        type: string
      status:
        description: 登录状态：0成功 1失败
        type: integer
      userName:
        description: 用户名
        type: string
    type: object
...
  github_com_gcc798_lightning_application_api_domain_response.Response:
    description: API 统一响应格式
    properties:
      code:
        description: 业务状态码
        example: 200
        type: integer
      data:
        description: 响应数据
        type: object
      msg:
        description: 响应消息
        example: 操作成功
        type: string
    type: object
...
host: localhost:9009
info:
  contact:
    email: support@example.com
    name: 技术支持
    url: https://example.com/support
  description: Quick Admin RESTful API 接口文档，支持双 Token 认证机制
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://example.com/terms/
  title: NTZ API 文档
  version: "1.0"
paths:
...
  /api/v1/loginLog:
    post:
      consumes:
      - application/json
      description: 创建新的登录日志记录
      parameters:
      - description: 创建登录日志请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_request.CreateLoginLogRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
      security:
      - Bearer: []
      summary: 创建登录日志
      tags:
      - 登录日志
    put:
      consumes:
      - application/json
      description: 更新登录日志记录
      parameters:
      - description: 更新登录日志请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_request.UpdateLoginLogRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
      security:
      - Bearer: []
      summary: 更新登录日志
      tags:
      - 登录日志
  /api/v1/loginLog/{id}:
    delete:
      consumes:
      - application/json
      description: 删除单个登录日志记录
      parameters:
      - description: 日志ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
      security:
      - Bearer: []
      summary: 删除登录日志
      tags:
      - 登录日志
    get:
      consumes:
      - application/json
      description: 根据日志ID查询登录日志详情
      parameters:
      - description: 日志ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.LoginLogResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
      security:
      - Bearer: []
      summary: 根据ID查询登录日志
      tags:
      - 登录日志
  /api/v1/loginLog/batch:
    delete:
      consumes:
      - application/json
      description: 批量删除登录日志记录
      parameters:
      - description: 批量删除请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_request.BatchDeleteLoginLogRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
      security:
      - Bearer: []
      summary: 批量删除登录日志
      tags:
      - 登录日志
  /api/v1/loginLog/clean:
    post:
      consumes:
      - application/json
      description: 清理指定天数之前的登录日志
      parameters:
      - description: 清理日志请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_request.CleanLoginLogRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 清理成功
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
      security:
      - Bearer: []
      summary: 清理登录日志
      tags:
      - 登录日志
...
  /api/v1/system/loginLog/page:
    post:
      consumes:
      - application/json
      description: 分页查询登录日志列表，支持按用户名、IP、状态、时间范围筛选
      parameters:
      - description: 页码
        in: query
        minimum: 1
        name: pageNum
        required: true
        type: integer
      - description: 每页数量
        in: query
        minimum: 1
        name: pageSize
        required: true
        type: integer
      - description: 排序列
        example: '"loginTime"'
        in: query
        name: orderByColumn
        type: string
      - description: 排序方向
        enum:
        - asc
        - desc
        example: '"desc"'
        in: query
        name: isAsc
        type: string
      - description: 用户名（模糊查询）
        in: query
        name: userName
        type: string
      - description: 登录IP（模糊查询）
        in: query
        name: ipaddr
        type: string
      - description: 登录状态：0成功 1失败，不传或null表示全部
        enum:
        - 0
        - 1
        in: query
        name: status
        type: integer
      - description: 开始时间
        example: '"2024-01-01 00:00:00"'
        in: query
        name: startTime
        type: string
      - description: 结束时间
        example: '"2024-12-31 23:59:59"'
        in: query
        name: endTime
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
      security:
      - Bearer: []
      summary: 分页查询登录日志列表
      tags:
      - 登录日志
...
securityDefinitions:
  Bearer:
    description: '格式: "Bearer {access_token}"，AccessToken 从登录接口获取'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...